Add final methods
